Refactor noop faas
This commit significantly refactors the noop FaaS so that
it can serve as a baseline for how all FaaS functions should
look.

Add a gman decorator to handle requesting new task IDs and
updating tasks on completion/delegation/failures. This will
allow the function writer to focus on functionality rather
than GMan housekeeping